chore(maple-core): drop redundant UInt cast in Hashing.BLAKE3.hex
ptr.count is already Int; the UInt(ptr.count) cast was redundant since
maple_blake3_hex accepts the platform's natural size type. No behaviour change.

+++ src/apple/Packages/MapleCore/Sources/MapleCore/Hashing.swift
@@ -27,7 +27,7 @@ public enum BLAKE3 {
             guard let base = ptr.baseAddress else { return -1 }
             return maple_blake3_hex(
                 base.assumingMemoryBound(to: UInt8.self),
-                UInt(ptr.count),
+                ptr.count,
                 &out)
         }
         guard rc == 0 else { return nil }
